+++ CTransformLine.h
@@ -11,13 +11,13 @@
 #include <map>
 #include "CPItem.h"
 class CTransformLine: public CPItem {
-  std::map<uint8_t, float> adc2c;
+  std::map<int16_t, float> adc2c;
   float result_offset;
 public:
   CTransformLine();
   virtual bool deSerialize(const JsonObject& root) override;
   virtual bool serialize(JsonObject& root) const override;
-  float convert(const uint8_t adc);
+  float convert(const int16_t adc);
 };
 
 #endif /* CTRANSFORMLINE_H_ */
